Break out sceAac from sceMp4.cpp/h
