+++ Core/Core.vcxproj
@@ -572,6 +572,7 @@
     <ClCompile Include="HLE\KUBridge.cpp" />
     <ClCompile Include="HLE\NetInetConstants.cpp" />
     <ClCompile Include="HLE\Plugins.cpp" />
+    <ClCompile Include="HLE\sceAac.cpp" />
     <ClCompile Include="HLE\sceKernelHeap.cpp" />
     <ClCompile Include="HLE\sceNetApctl.cpp" />
     <ClCompile Include="HLE\sceNetInet.cpp" />
@@ -1191,6 +1192,7 @@
     <ClInclude Include="HLE\KUBridge.h" />
     <ClInclude Include="HLE\NetInetConstants.h" />
     <ClInclude Include="HLE\Plugins.h" />
+    <ClInclude Include="HLE\sceAac.h" />
     <ClInclude Include="HLE\sceKernelHeap.h" />
     <ClInclude Include="HLE\sceNetApctl.h" />
     <ClInclude Include="HLE\sceNetInet.h" />
