+++ Core/Core.vcxproj.filters
@@ -1339,6 +1339,9 @@
     <ClCompile Include="HLE\SocketManager.cpp">
       <Filter>HLE\Libraries</Filter>
     </ClCompile>
+    <ClCompile Include="HLE\sceAac.cpp">
+      <Filter>HLE\Libraries</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="ELF\ElfReader.h">
@@ -2155,6 +2158,9 @@
       <Filter>HLE\Libraries</Filter>
     </ClInclude>
     <ClInclude Include="HLE\SocketManager.h">
+      <Filter>HLE\Libraries</Filter>
+    </ClInclude>
+    <ClInclude Include="HLE\sceAac.h">
       <Filter>HLE\Libraries</Filter>
     </ClInclude>
   </ItemGroup>
