+++ Core/HLE/sceMp4.cpp
@@ -25,29 +25,6 @@
 #include "Core/HLE/sceMp4.h"
 #include "Core/HW/SimpleAudioDec.h"
 
-static std::map<u32, AuCtx*> aacMap;
-
-static AuCtx *getAacCtx(u32 id) {
-	if (aacMap.find(id) == aacMap.end())
-		return NULL;
-	return aacMap[id];
-}
-
-void __AACShutdown() {
-	for (auto it = aacMap.begin(), end = aacMap.end(); it != end; it++) {
-		delete it->second;
-	}
-	aacMap.clear();
-}
-
-void __AACDoState(PointerWrap &p) {
-	auto s = p.Section("sceAAC", 0, 1);
-	if (!s)
-		return;
-
-	Do(p, aacMap);
-}
-
 static u32 sceMp4Init()
 {
 	INFO_LOG(Log::ME, "sceMp4Init()");
@@ -194,207 +171,6 @@ static u32 sceMp4SearchSyncSampleNum()
 }
 
 
-// sceAac module starts from here
-
-static u32 sceAacExit(u32 id)
-{
-	INFO_LOG(Log::ME, "sceAacExit(id %i)", id);
-	if (aacMap.find(id) != aacMap.end()) {
-		delete aacMap[id];
-		aacMap.erase(id);
-	}
-	else{
-		ERROR_LOG(Log::ME, "%s: bad aac id %08x", __FUNCTION__, id);
-		return -1;
-	}
-	return 0;
-}
-
-static u32 sceAacInit(u32 id)
-{
-	INFO_LOG(Log::ME, "UNIMPL sceAacInit(%08x)", id);
-	if (!Memory::IsValidAddress(id)){
-		ERROR_LOG(Log::ME, "sceAacInit() AAC Invalid id address %08x", id);
-		return ERROR_AAC_INVALID_ADDRESS;
-	}
-
-	AuCtx *aac = new AuCtx();
-	aac->startPos = Memory::Read_U64(id);				// AUDIO stream start position.
-	aac->endPos = Memory::Read_U32(id + 8);				// AUDIO stream end position.
-	aac->AuBuf = Memory::Read_U32(id + 16);             // Input AAC data buffer.	
-	aac->AuBufSize = Memory::Read_U32(id + 20);         // Input AAC data buffer size.
-	aac->PCMBuf = Memory::Read_U32(id + 24);            // Output PCM data buffer.
-	aac->PCMBufSize = Memory::Read_U32(id + 28);        // Output PCM data buffer size.
-	aac->freq = Memory::Read_U32(id + 32);              // Frequency.
-	if (aac->AuBuf == 0 || aac->PCMBuf == 0) {
-		ERROR_LOG(Log::ME, "sceAacInit() AAC INVALID ADDRESS AuBuf %08x PCMBuf %08x", aac->AuBuf, aac->PCMBuf);
-		delete aac;
-		return ERROR_AAC_INVALID_ADDRESS;
-	}
-	if (aac->startPos > aac->endPos) {
-		ERROR_LOG(Log::ME, "sceAacInit() AAC INVALID startPos %lli endPos %lli", aac->startPos, aac->endPos);
-		delete aac;
-		return ERROR_AAC_INVALID_PARAMETER;
-	}
-	if (aac->AuBufSize < 8192 || aac->PCMBufSize < 8192) {
-		ERROR_LOG(Log::ME, "sceAacInit() AAC INVALID PARAMETER, bufferSize %i outputSize %i", aac->AuBufSize, aac->PCMBufSize);
-		delete aac; 
-		return ERROR_AAC_INVALID_PARAMETER;
-	}
-	if (aac->freq != 24000 && aac->freq != 32000 && aac->freq != 44100 && aac->freq != 48000) {
-		ERROR_LOG(Log::ME, "sceAacInit() AAC INVALID freq %i", aac->freq);
-		delete aac;
-		return ERROR_AAC_INVALID_PARAMETER;
-	}
-
-	DEBUG_LOG(Log::ME, "startPos %llx endPos %llx AuBuf %08x AuBufSize %08x PCMbuf %08x PCMbufSize %08x freq %d",
-		aac->startPos, aac->endPos, aac->AuBuf, aac->AuBufSize, aac->PCMBuf, aac->PCMBufSize, aac->freq);
-
-	aac->Channels = 2;
-	aac->MaxOutputSample = aac->PCMBufSize / 4;
-	aac->SetReadPos((int)aac->startPos);
-
-	// create aac decoder
-	aac->decoder = CreateAudioDecoder(PSP_CODEC_AAC);
-
-	// close the audio if id already exist.
-	if (aacMap.find(id) != aacMap.end()) {
-		delete aacMap[id];
-		aacMap.erase(id);
-	}
-	aacMap[id] = aac;
-
-	return id;
-}
-
-static u32 sceAacInitResource(u32 numberIds)
-{
-	// Do nothing here
-	INFO_LOG_REPORT(Log::ME, "sceAacInitResource(%i)", numberIds);
-	return 0;
-}
-
-static u32 sceAacTermResource()
-{
-	ERROR_LOG(Log::ME, "UNIMPL sceAacTermResource()");
-	return 0;
-}
-
-static u32 sceAacDecode(u32 id, u32 pcmAddr)
-{
-	// return the size of output pcm, <0 error
-	DEBUG_LOG(Log::ME, "sceAacDecode(id %i, bufferAddress %08x)", id, pcmAddr);
-	auto ctx = getAacCtx(id);
-	if (!ctx) {
-		ERROR_LOG(Log::ME, "%s: bad aac id %08x", __FUNCTION__, id);
-		return -1;
-	}
-
-	return ctx->AuDecode(pcmAddr);
-}
-
-static u32 sceAacGetLoopNum(u32 id)
-{
-	INFO_LOG(Log::ME, "sceAacGetLoopNum(id %i)", id);
-	auto ctx = getAacCtx(id);
-	if (!ctx) {
-		ERROR_LOG(Log::ME, "%s: bad aac id %08x", __FUNCTION__, id);
-		return -1;
-	}
-	return ctx->LoopNum;
-}
-
-static u32 sceAacSetLoopNum(u32 id, int loop)
-{
-	INFO_LOG(Log::ME, "sceAacSetLoopNum(id %i,loop %d)", id, loop);
-	auto ctx = getAacCtx(id);
-	if (!ctx) {
-		ERROR_LOG(Log::ME, "%s: bad aac id %08x", __FUNCTION__, id);
-		return -1;
-	}
-
-	ctx->LoopNum = loop;
-	return 0;
-}
-
-static int sceAacCheckStreamDataNeeded(u32 id)
-{
-	// return 1 to read more data stream, 0 don't read, <0 error
-	DEBUG_LOG(Log::ME, "sceAacCheckStreamDataNeeded(%i)", id);
-
-	auto ctx = getAacCtx(id);
-	if (!ctx) {
-		ERROR_LOG(Log::ME, "%s: bad aac id %08x", __FUNCTION__, id);
-		return -1;
-	}
-
-	return ctx->AuCheckStreamDataNeeded();
-}
-
-static u32 sceAacNotifyAddStreamData(u32 id, int size)
-{
-	// check how many bytes we have read from source file
-	DEBUG_LOG(Log::ME, "sceAacNotifyAddStreamData(%i, %08x)", id, size);
-
-	auto ctx = getAacCtx(id);
-	if (!ctx) {
-		ERROR_LOG(Log::ME, "%s: bad aac id %08x", __FUNCTION__, id);
-		return -1;
-	}
-
-	return ctx->AuNotifyAddStreamData(size);
-}
-
-static u32 sceAacGetInfoToAddStreamData(u32 id, u32 buff, u32 size, u32 srcPos)
-{
-	// read from stream position srcPos of size bytes into buff
-	DEBUG_LOG(Log::ME, "sceAacGetInfoToAddStreamData(%08X, %08X, %08X, %08X)", id, buff, size, srcPos);
-
-	auto ctx = getAacCtx(id);
-	if (!ctx) {
-		ERROR_LOG(Log::ME, "%s: bad aac handle %08x", __FUNCTION__, id);
-		return -1;
-	}
-
-	return ctx->AuGetInfoToAddStreamData(buff, size, srcPos);
-}
-
-static u32 sceAacGetMaxOutputSample(u32 id)
-{
-	DEBUG_LOG(Log::ME, "sceAacGetMaxOutputSample(id %i)", id);
-	auto ctx = getAacCtx(id);
-	if (!ctx) {
-		ERROR_LOG(Log::ME, "%s: bad aac id %08x", __FUNCTION__, id);
-		return -1;
-	}
-
-	return ctx->MaxOutputSample;
-}
-
-static u32 sceAacGetSumDecodedSample(u32 id)
-{
-	DEBUG_LOG(Log::ME, "sceAacGetSumDecodedSample(id %i)", id);
-	auto ctx = getAacCtx(id);
-	if (!ctx) {
-		ERROR_LOG(Log::ME, "%s: bad aac id %08x", __FUNCTION__, id);
-		return -1;
-	}
-
-	return ctx->SumDecodedSamples;
-}
-
-static u32 sceAacResetPlayPosition(u32 id)
-{
-	INFO_LOG(Log::ME, "sceAacResetPlayPosition(id %i)", id);
-	auto ctx = getAacCtx(id);
-	if (!ctx) {
-		ERROR_LOG(Log::ME, "%s: bad aac id %08x", __FUNCTION__, id);
-		return -1;
-	}
-
-	return ctx->AuResetPlayPosition();
-}
-
 static u32 mp4msv_3C2183C7(u32 unknown1, u32 unknown2) {
 	ERROR_LOG(Log::ME, "UNIMPL mp4msv_3C2183C7(%d, %x)", unknown1, unknown2);
 	return 0;
@@ -448,23 +224,6 @@ const HLEFunction sceMp4[] =
 	{0X9CA13D1A, nullptr,                                        "mp4msv_9CA13D1A",                   '?', ""       },
 };
 
-// 395
-const HLEFunction sceAac[] = {
-	{0XE0C89ACA, &WrapU_U<sceAacInit>,                           "sceAacInit",                        'x', "x"      },
-	{0X33B8C009, &WrapU_U<sceAacExit>,                           "sceAacExit",                        'x', "x"      },
-	{0X5CFFC57C, &WrapU_U<sceAacInitResource>,                   "sceAacInitResource",                'x', "x"      },
-	{0X23D35CAE, &WrapU_V<sceAacTermResource>,                   "sceAacTermResource",                'x', ""       },
-	{0X7E4CFEE4, &WrapU_UU<sceAacDecode>,                        "sceAacDecode",                      'x', "xx"     },
-	{0X523347D9, &WrapU_U<sceAacGetLoopNum>,                     "sceAacGetLoopNum",                  'x', "x"      },
-	{0XBBDD6403, &WrapU_UI<sceAacSetLoopNum>,                    "sceAacSetLoopNum",                  'x', "xi"     },
-	{0XD7C51541, &WrapI_U<sceAacCheckStreamDataNeeded>,          "sceAacCheckStreamDataNeeded",       'i', "x"      },
-	{0XAC6DCBE3, &WrapU_UI<sceAacNotifyAddStreamData>,           "sceAacNotifyAddStreamData",         'x', "xi"     },
-	{0X02098C69, &WrapU_UUUU<sceAacGetInfoToAddStreamData>,      "sceAacGetInfoToAddStreamData",      'x', "xxxx"   },
-	{0X6DC7758A, &WrapU_U<sceAacGetMaxOutputSample>,             "sceAacGetMaxOutputSample",          'x', "x"      },
-	{0X506BF66C, &WrapU_U<sceAacGetSumDecodedSample>,            "sceAacGetSumDecodedSample",         'x', "x"      },
-	{0XD2DA2BBA, &WrapU_U<sceAacResetPlayPosition>,              "sceAacResetPlayPosition",           'x', "x"      },
-};
-
 const HLEFunction mp4msv[] = {
 	{0x3C2183C7, &WrapU_UU<mp4msv_3C2183C7>,                    "mp4msv_3C2183C7",               'x', "xx"      },
 	{0x9CA13D1A, &WrapU_UU<mp4msv_9CA13D1A>,                    "mp4msv_9CA13D1A",               'x', "xx"      },
@@ -474,7 +233,6 @@ const HLEFunction mp4msv[] = {
 void Register_sceMp4()
 {
 	RegisterModule("sceMp4", ARRAY_SIZE(sceMp4), sceMp4);
-	RegisterModule("sceAac", ARRAY_SIZE(sceAac), sceAac);	
 }
 
 void Register_mp4msv()
