+++ Core/HLE/sceMp4.h
@@ -17,12 +17,5 @@
 
 #pragma once
 
-enum {
-	ERROR_AAC_INVALID_ADDRESS = 0x80691002,
-	ERROR_AAC_INVALID_PARAMETER = 0x80691003,
-};
-
 void Register_sceMp4();
-void __AACShutdown();
-void __AACDoState(PointerWrap &p);
 void Register_mp4msv();
