+++ UWP/CoreUWP/CoreUWP.vcxproj
@@ -231,6 +231,7 @@
     <ClInclude Include="..\..\Core\HLE\sceMd5.h" />
     <ClInclude Include="..\..\Core\HLE\sceMp3.h" />
     <ClInclude Include="..\..\Core\HLE\sceMp4.h" />
+    <ClInclude Include="..\..\Core\HLE\sceAac.h" />
     <ClInclude Include="..\..\Core\HLE\sceMpeg.h" />
     <ClInclude Include="..\..\Core\HLE\sceMt19937.h" />
     <ClInclude Include="..\..\Core\HLE\sceNet.h" />
@@ -491,6 +492,7 @@
     <ClCompile Include="..\..\Core\HLE\sceMd5.cpp" />
     <ClCompile Include="..\..\Core\HLE\sceMp3.cpp" />
     <ClCompile Include="..\..\Core\HLE\sceMp4.cpp" />
+    <ClCompile Include="..\..\Core\HLE\sceAac.cpp" />
     <ClCompile Include="..\..\Core\HLE\sceMpeg.cpp" />
     <ClCompile Include="..\..\Core\HLE\sceMt19937.cpp" />
     <ClCompile Include="..\..\Core\HLE\sceNet.cpp" />
